+++ src/conspiracies/__init__.py
@@ -1,2 +1,5 @@
-from .component_HeadWordExtraction import create_headwords_component  # noqa
-from .coref import create_coref_component  # noqa
+from .HeadWordExtractionComponent import (  # noqa
+    create_headwords_component,
+    HeadwordsExtractionComponent,
+)
+from .coref import create_coref_component, CoreferenceComponent  # noqa
