+++ src/conspiracies/about.py
@@ -1,3 +1,3 @@
-__version__ = "0.0.2"
+__version__ = "0.0.3"
 __title__ = "conspiracies"
 __download_url__ = "https://github.com/centre-for-humanities-computing/conspiracies"
