+++ src/conspiracies/coref/CoreferenceComponent.py
