+++ src/conspiracies/coref/__init__.py
@@ -1,2 +1,2 @@
 from .CoreferenceModel import CoreferenceModel  # noqa
-from .component import CoreferenceComponent, create_coref_component  # noqa
+from .CoreferenceComponent import CoreferenceComponent, create_coref_component  # noqa
